Chg per PMcManus on lawful intercept
Per PMM:
I appreciate the authors kicking off the effort with this draft that proposes phase 2 requirements.

On a couple of occasions the draft makes creates requirements wrt law enforcement compliance. e.g. "comply with locally relevant law enforcement [..] (high priority)"..

While RFC 2804 says
"The IETF has decided not to consider requirements for wiretapping as part of the process for creating and maintaining IETF standards.". There is clear guidance for avoiding this requirement in any document designed to create an expectation of privacy from end to end, before it is adopted by a WG.

-Patrick

JL: Generalized to comply with applicable legal requirements